Written (most of) Python script for updating README contents.
A Python script has been added to the project that will be responsible for calling the GitHub events API and formatting its contents for insertion into the README file, the latter of which has also been cleaned up and prepared for work from the action once it is set up.

Because of the complexity of some event types returned by the API, not all events have been accounted for yet, which I have decided to sort out at a later date once the action is up and running.